+++ packages/discovery-react-components/src/components/DocumentPreview/DocumentPreview.tsx
@@ -9,6 +9,7 @@ import SimpleDocument from './components/SimpleDocument/SimpleDocument';
 import withErrorBoundary, { WithErrorBoundaryProps } from 'utils/hoc/withErrorBoundary';
 import { defaultMessages, Messages } from './messages';
 import HtmlView from './components/HtmlView/HtmlView';
+import PdfViewerWithHighlight from './components/PdfViewerHighlight/PdfViewerWithHighlight';
 import { getTextMappings, isCsvFile, isJsonFile } from './utils/documentData';
 import { TextMappings } from './types';
 
@@ -230,6 +231,7 @@ function PreviewDocument({
 const ErrorBoundDocumentPreview: any = withErrorBoundary(DocumentPreview);
 ErrorBoundDocumentPreview.PreviewToolbar = PreviewToolbar;
 ErrorBoundDocumentPreview.PreviewDocument = PreviewDocument;
+ErrorBoundDocumentPreview.PdfViewerWithHighlight = PdfViewerWithHighlight;
 
 export default ErrorBoundDocumentPreview;
 export { ErrorBoundDocumentPreview as DocumentPreview };