+++ examples/discovery-search-app/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0-beta.7](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.6...v1.5.0-beta.7) (2021-12-13)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+# [1.5.0-beta.6](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.5...v1.5.0-beta.6) (2021-12-10)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+# [1.5.0-beta.5](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.4...v1.5.0-beta.5) (2021-12-09)
+
+
+### Features
+
+* add PDF viewer with highlighting ([#238](https://github.com/watson-developer-cloud/discovery-components/issues/238)) ([5e06d62](https://github.com/watson-developer-cloud/discovery-components/commit/5e06d62ac146e202c4a7c2a1af5bca6258d746c8))
+
+
+
+
+
 # [1.5.0-beta.4](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.3...v1.5.0-beta.4) (2021-12-06)
 
 **Note:** Version bump only for package discovery-search-app